Add disconnect-all-of-group menu item to patchbay
If one group has multiple connections to another group an extra menu
item is added to its disconnect context menu, which allows the user to
disconnect it all at once. The entry is added at the bottom of the
disconnect menu in the form of "Groupname (all x ports)". This comes in
very handy when the user wants to disconnect a soundcard from their
current application.

This commit changes the API of PortContextMenuDisconnect(). Data is now
passed as an argument to this method instead of transported via QT's
setData()/data(). This has been done to allow us to disconnect multiple
connections at once.